+++ pkgs/jnigen/CHANGELOG.md
@@ -1,6 +1,8 @@
 ## 0.15.1-wip
 
 - Add docs about debugging.
+- Add support for Kotlin interfaces with suspend functions. These can now be
+  implemented using Dart functions that return a `Future`.
 
 ## 0.15.0